Various RDF handling improvements and refactorization.

+++ URSA.Example.WebApplication/Data/IProduct.cs
@@ -19,7 +19,7 @@ namespace URSA.Example.WebApplication.Data
 
         /// <summary> Gets or sets the products's price.</summary>
         [Property("http://temp.uri/vocab#price")]
-        float Price { get; set; }
+        double Price { get; set; }
 
         /// <summary> Gets or sets the products's features.</summary>
         [Collection("http://temp.uri/vocab#feature")]
